requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==2.0.2
pandas==2.3.3
py4j==0.10.9.9
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
